IdentityUser and settings of Roles for a Administrator now runs perfectly. The Administrator can now add a user and it saves correctly to the Db. Only the Administrator will be able to see the button and div for Add User.
