+++ KingPim.Web/Controllers/HomeController.cs
@@ -1,21 +1,63 @@
-﻿using KingPim.Repositories.Interfaces;
+﻿using KingPim.Models.ViewModels;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using System.Threading.Tasks;
 
 namespace KingPim.Web.Controllers
 {
     [Authorize]
     public class HomeController : Controller
     {
-        private IHome homeRepo;
+        private readonly UserManager<IdentityUser> _userManager;
+        private readonly SignInManager<IdentityUser> _signInManager;
+        private readonly RoleManager<IdentityRole> _roleManager;
 
-        public HomeController(IHome homeRepository)
+        public HomeController(UserManager<IdentityUser> userManager, SignInManager<IdentityUser> signInManager, RoleManager<IdentityRole> roleManager)
         {
-            homeRepo = homeRepository;
+            _userManager = userManager;
+            _signInManager = signInManager;
+            _roleManager = roleManager;
         }
+        [AllowAnonymous]
         public IActionResult Index()
         {
-            return View();
+            // Checks if the user is authenticated/signed in and redirectsthe user to Account: 
+            if (User.Identity.IsAuthenticated)
+            {
+                return RedirectToAction("Index", "Account");
+            }
+            else
+            {
+                // Otherwise, maybe return to an error page?
+                return View();
+            }
+        }
+
+        // Checks if the password matches to the account, redirects the user to Account:
+        [AllowAnonymous]
+        public async Task<IActionResult> Login(LoginViewModel vm)
+        {
+            if (ModelState.IsValid)
+            {
+                var user = await _userManager.FindByEmailAsync(vm.UserName);
+                if (user != null)
+                {
+                    await _signInManager.SignOutAsync();
+                    if ((await _signInManager.PasswordSignInAsync(user, vm.Password, false, false)).Succeeded)
+                    {
+                        return RedirectToAction("Index", "Account");
+                    }
+                }
+            }
+            return View("Index", vm);
+        }
+
+        // Sends the user back to the login page:
+        public async Task<IActionResult> Logout()
+        {
+            await _signInManager.SignOutAsync();
+            return RedirectToAction(nameof(Index));
         }
     }
 }
