+++ KingPim.Web/Views/Account/Index.cshtml
@@ -1,23 +1,34 @@
-﻿@model LoginViewModel
+﻿@model AccountViewModel
 
-@{
-    Layout = "~/Views/Shared/_LoginLayout.cshtml";
-}
+<div class="container">
+    <h1 class="text-center" style="margin-bottom:50px;"><strong>Welcome to KingPim!</strong></h1>
+    <div class="row">
+        <div class="col-12" style="background-color:lightgrey;margin-bottom:100px;border-radius:10px;">
+            <h3 class="text-center" style="margin-top:10px"><strong>Create Account</strong></h3>
+            <form asp-controller="Account" asp-action="Add" method="post">
+                <div class="form-group">
+                    <label asp-for="@Model.UserName"><strong></strong></label>
+                    <input asp-for="@Model.UserName" placeholder="UserName..." class="form-control form-control-sm" />
 
-<div class="container mx-auto" style="width: 330px; margin-top: 50px">
-    <form asp-controller="Account" asp-action="Login" method="post" style="max-width: 330px;">
-        <h2 class="form-login-heading" style="text-align: center; margin: 10px"><strong>Please sign in</strong></h2>
-        <label asp-for="UserName" class="sr-only">Email address</label>
-        <input asp-for="UserName" id="inputUserName" class="form-control" placeholder="Email address" required autofocus style="margin:10px" />
-        <div><span asp-validation-for="UserName" class="text-danger"></span></div>
-        <label asp-for="Password" class="sr-only">Password</label>
-        <input asp-for="Password" id="inputPassword" class="form-control" placeholder="Password" required style="margin:10px" />
-        <div><span asp-validation-for="Password" class="text-danger"></span></div>
-        <button class="btn btn-lg btn-primary btn-block" type="submit" style="margin:10px">Sign in</button>
-    </form>
-</div>
-<div class="container mx-auto" style="width: 330px;">
-    <div class="text-center">
-        <a style="color:red; margin-right:5px;"><strong>Forgotten password?</strong></a>
+                    <label asp-for="@Model.Password"><strong></strong></label>
+                    <input asp-for="@Model.Password" placeholder="Password..." class="form-control form-control-sm" />
+
+                    <label asp-for="@Model.ConfirmPassword"><strong></strong></label>
+                    <input asp-for="@Model.ConfirmPassword" placeholder="Confirm Password..." class="form-control form-control-sm" />
+                    <br />
+                    <div class="form-group">
+                        <select class="form-control" name="Roles">
+                            <option value="">Select a role</option>
+                            <option value="Administrator">Administrator</option>
+                            <option value="Editor">Editor</option>
+                            <option value="Publisher">Publisher</option>
+                        </select>
+                    </div>
+                </div>
+                <div class="text-center" style="margin-bottom:50px;margin-top:100px;">
+                    <button type="submit" id="createAccount" class="btn btn-success">Create</button>
+                </div>
+            </form>
+        </div>
     </div>
 </div>
