+++ KingPim.Web/Views/Home/Index.cshtml
@@ -1,46 +1,23 @@
-﻿@model AccountViewModel
+﻿@model LoginViewModel
 
-    <div class="container">
-        <h1 class="text-center" style="margin-bottom:50px;"><strong>Welcome to KingPim!</strong></h1>
-        <div class="row">
-            <div class="col-6" style="background-color:lightgrey;padding-bottom:100px;border-radius:10px;">
-                <h3 class="text-center" style="margin-top:10px"><strong>Create Account</strong></h3>
-                <form asp-controller="Account" asp-action="Index" method="post">
-                    <div class="form-group">
-                            <label asp-for="@Model.UserName"><strong></strong></label>
-                            <input asp-for="@Model.UserName" placeholder="UserName..." class="form-control form-control-sm" />
+@{
+    Layout = "~/Views/Shared/_LoginLayout.cshtml";
+}
 
-                            <label asp-for="@Model.Password"><strong></strong></label>
-                            <input asp-for="@Model.Password" placeholder="Password..." class="form-control form-control-sm" />
-
-                            <label asp-for="@Model.ConfirmPassword"><strong></strong></label>
-                            <input asp-for="@Model.ConfirmPassword" placeholder="Confirm Password..." class="form-control form-control-sm" />
-                                <select @*asp-for=""*@ class="form-control-sm" required style="margin-top:10px">
-                                    <option value="" selected disabled>Select a role</option>
-                                    <option>Administrator</option>
-                                    <option>Publisher</option>
-                                    <option>Editor</option>
-                                </select>
-                                @*<div class="form-group">
-                                    <select asp-for="RolesId" class="form-control" required>
-                                        <option value="" selected>Select a subcategory</option>
-                                        @foreach (var r in Model.Roles)
-                                        {
-                                            <option value="@r.Id">@r.Name</option>
-                                        }
-                                    </select>
-                                    <br />
-                                </div>*@
-                    </div>
-                    <div class="text-center">
-                        <button type="button" class="btn btn-danger">Close</button>
-                        <input type="hidden" name="categoryId" value="" />
-                        <button type="submit" id="createAccount" class="btn btn-success">Create</button>
-                    </div>
-                </form>
-            </div>
-            @*<div class="col-6" style="background-color:lightgray;padding-bottom:300px">
-                <h3 class="text-center" style="margin:10px"><strong>All the accounts</strong></h3>
-            </div>*@
-        </div>
+<div class="container mx-auto" style="width: 330px; margin-top: 50px">
+    <form asp-controller="Home" asp-action="Login" method="post" style="max-width: 330px;">
+        <h2 class="form-login-heading" style="text-align: center; margin: 10px"><strong>Please sign in</strong></h2>
+        <label asp-for="UserName" class="sr-only">Email address</label>
+        <input asp-for="UserName" id="inputUserName" class="form-control" placeholder="Email address" required autofocus style="margin:10px" />
+        <div><span asp-validation-for="UserName" class="text-danger"></span></div>
+        <label asp-for="Password" class="sr-only">Password</label>
+        <input asp-for="Password" id="inputPassword" class="form-control" placeholder="Password" required style="margin:10px" />
+        <div><span asp-validation-for="Password" class="text-danger"></span></div>
+        <button class="btn btn-lg btn-primary btn-block" type="submit" style="margin:10px">Sign in</button>
+    </form>
+</div>
+<div class="container mx-auto" style="width: 330px;">
+    <div class="text-center">
+        <a style="color:red; margin-right:5px;"><strong>Forgotten password?</strong></a>
     </div>
+</div>
